Update devScatric branch to the last of devNako

+++ db.py
@@ -13,7 +13,7 @@ def getDB():
 def getUser(user_id):
     cursor = getDB().cursor()
     if user_id.isnumeric():
-        curs = cursor.execute(f'SELECT * FROM users WHERE id={user_id}')
+        curs = cursor.execute(f"SELECT * FROM users WHERE id={user_id};")
         return curs.fetchone()
     else:
         try: curs = cursor.execute(f"SELECT * FROM users WHERE username='{user_id}'")
@@ -39,7 +39,7 @@ def deleteUser(user_id):
     cursor = getDB().cursor()
     if user_id.isnumeric():
         try:                
-            curs = cursor.execute(f"DELETE FROM users WHERE id='{user_id}'")
+            curs = cursor.execute(f"DELETE FROM users WHERE id = {user_id}; DELETE FROM posts WHERE user_id = {user_id}")
             g.db.commit()
         except sqlite3.Error as er:
             print(f"[{datetime.datetime.now()}]: {er.sqlite_errorname} | {er} | {er.sqlite_errorcode}")
@@ -47,10 +47,32 @@ def deleteUser(user_id):
             else: return 'error'
         else: return 'success'
 
-def createPost(content):
+def createPost(content, user_id):
     cursor = getDB().cursor()
-    curs = cursor.execute(f"INSERT INTO posts (content) VALUES ('{content}');")
+    curs = cursor.execute(f"INSERT INTO posts (content, user_id) VALUES ('{content}', {user_id});")
     curs.commit()
+
+def deletePost(post_id):
+    cursor = getDB().cursor()
+    curs = cursor.execute(f"DELETE FROM posts WHERE post_id = {post_id};")
+    curs.commit()
+
+def getPost(post_id):
+    cursor = getDB().cursor()
+    curs = cursor.execute(f"SELECT * FROM posts WHERE post_id = {post_id};")
+    return curs.fetchone()
+
+def changeUser(user_id, username):
+    cursor = getDB().cursor()
+    curs = cursor.execute(f"UPDATE users SET username = '{username}' WHERE id = {user_id};")
+    curs.commit()
+
+def modifyPost(post_id, content):
+    cursor = getDB().cursor()
+    curs = cursor.execute(f"UPDATE posts SET content = '{content}' WHERE post_id = {post_id};")
+    curs.commit()
+
+#To Do: Show posts(user_id)
 
 @app.teardown_appcontext
 def teardown_db(exception):
